+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 
 - Initial version, proof of concept.
 - Very little checking of input at this stage.
+
+## 0.1.1
+
+- Basic numerical differentiation
+- Basic numerical integration (Simpson)
+- Basic zero-finding (Newton)